chore: moved theme-provider to the other providers
