+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import "~/styles/globals.css";
-import { ThemeProvider } from "~/components/theme-provider";
+import { ThemeProvider } from "~/components/providers/theme-provider";
 import type { Metadata, Viewport } from "next";
 import { cn } from "~/lib/utils";
 import { ClerkProvider } from "@clerk/nextjs";
